Only treat "OK" as a response terminator when appropriate
Right now __read() will stop reading at "OK" even if a ">" is still coming.
This can cause the ">" to be seen as the response to the next command, which
confuses the initialization sequence, since the initialization sequence
expects a very specific set of responses to its commands.

This changes __read() so that by default it only treats ">" as the response
terminator. When we issue the "ATLP" command to enter low-power mode, we will
use "OK" as the response terminator instead, since that's the only time we
don't expect to see a prompt.

This should fix #226 and should also fix #227.